Introduce GMD + ATD
Introduce Gradle Managed Device definitions to modules that have instrumented tests.

At the moment, defining it as a convention plugin isn't possible due to the bug. That bug is now fixed and will be included in Android Studio Flamingo.
Once Flamingo becomes stable, convert the GMD definitions as a convention plugin

Change-Id: I9e22ee2ec1f8a9530292b42de6f661561034c350